Add correctness validation for more categories
There is actually no reason not to check correctness for nooverflows
and memsafety. uautomizer does this, so it is good to do this as well.
uautomizer was also best at verifying nooverflow last year, therefore
we choose it as verification backend for nooverflow correctness.
For memsafety we use symbiotic because of the same reason.